Add Stardate script
This script is used across various systems in:
- screen hardstatus line
- tmux status line

The script provides a simplistic Stardate calculator, not based
on any particular methodology of the Star Trek series, but on
a simple calculation based off GMT time.

The firt 2 digits are the number of years since 1970, the next
three digits and the digit after the decimal point indicate the
"stardate" of the year, which is a straightforward multiplication
by ~2.73785 which is 1000 stardates per calendar year which is an
average of 365.25 days.