Tweak sdate script to print correct Julian date
The existing script was rounding the computed Julian date which was
resulting in the script printing the next day for 12 hours in the day,
i.e., a date of 456314.5 was being printed as 45631.5 (due to the
division by 10 and rounding up by the print statement). Now, the script
computes the number of days since the epoch (JD 2000000) and drops the
fractional portion, so it will print the correct Julian date stardate,
i.e., a date of 456314.5 will be printed correctly as 45631.4.