Adjust status bar colors in tmux
With Terminal.app, the background color was getting set to base02
instead of base03, therefore, I had to use base03 (brightblack) to
distinguish the tmux status bar from the rest of the screen. It would
be a problem with vim (not so much really, since I could figure out
where the tmux status bar was based on the vim status line).

However, since iTerm2 is a better terminal emulator, it correctly
displays the background color as base03, so in order to distinguish the
status bar, I changed the color to base02 (black) instead.