Add commands to set executable permission on file
The deploy script was simply redirecting the contents of the files into
the target files, however, they were getting created without executable
permissions. Since this script is designed to deploy executables, it is
required to set the executable permissions.

This cannot be done in the deployment configuration file however,
because the script parses all the input files first before extracting
the relevant contents. Therefore, if we try to run chmod +x within the
file, it would fail because the file would not exist yet (or run deploy
twice).

This commit also cleans up the formatting when building the command line
so that each command is set up on it's own line.