Add Git commit template
This configures Git to use a standard commit template, in order to make
it easier to write better commit messages.

This commit adds a default template, and configures Git to use it as the
default. It also modifies the install script to install the default
template at ~/.git-commit-template. Finally, this adds an extra check to
reapply the committed Git configuration to the global Git config.