Split prompt into two lines
Prior to this change, we were setting PS1 manually and having to muck
with the \[, \] pairs for terminal escapes was proving to be too much.
Instead, we moved to using a function used in PROMPT_COMMAND which
displays the user, hostname and directory on one line, followed by the
actual prompt line which displays the git status.