Enable default mouse option in tmux
Prior to this change, I was using older versions of tmux. However, with
recent changes, I have got access to at least tmux version 2.0, which
removes the old mouse-* options and replaces them with a single `mouse`
option.

This change removes the settings for the old mouse-* options and
replaces them with the single mouse option.