Add Zsh config files
Prior to this change, I was using Bash as the default shell, but I have
now started using Zsh, which has more features.

This change adds the zshrc and related plugin files to source control
and installs them along with the bashrc files. This also updates tmux to
use zsh as the default shell and command.