Move alias for ls into lscolors.zsh
Having the ls alias default to `ls --color=auto` caused it to break on
OSX, since the version of ls that ships with OSX doesn't support the
`--color` option.

This change moves the alias into lscolors, since we already have a check
for OSX vs non-OSX (in my case, Linux). This allows the alias to work as
expected on Linux, while the environment variable LSCOLORS provides the
coloring option on OSX.