Always (re)set LS_COLORS
Prior to this change, if I changed the dircolors configuration and
restarted the shell, it wouldn't reflect in the new shell because of the
check. I'd have to create a new terminal window to get the new colors.

This change eliminates the check for a set LS_COLORS variable and always
calls dircolors.