Add virtualenv to prompt, if it exists
With my starting to get deeper into Python development, I noticed an
issue with the virtualenv, where the current virtualenv was not set on
activation. I traced this to the fact that PS1 is always reset in the
prompt_header function, therefore, I need to add the virtual environment
details into prompt_header

This change checks if the virtual environment is set, if so, it adds the
current virtual environment to the prompt in magenta color.